requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
celery==5.4.0
django-prometheus
pandas==2.2.2 
xlsxwriter==3.2.0
